+++ Packages/com.unity.ide.vscode/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Code Editor Package for Visual Studio Code
 
+## [1.1.1] - 2019-08-19
+
+Support for Player Project. Generates specific csproj files containing files, reference, defines,
+etc. that will show how the assembly will be compiled for a target platform.
+
+
 ## [1.1.0] - 2019-08-07
 
 Adds support for choosing extensions to be opened with VSCode. This can be done through the GUI in Preferences.